changed folder structure and named renamed folders using PascalCase
