+++ client/src/components/DetailDataBoxParts/characters.jsx
