+++ client/src/components/DetailDataBoxParts/mainInfo.jsx
