+++ client/src/components/DetailDataBoxParts/settings.jsx
