+++ client/src/index.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-undef */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/app.jsx';
+import App from './app.jsx';
 
 ReactDOM.render(<App />, document.getElementById('app'));
