+++ package.json
@@ -8,10 +8,10 @@
     "node": ">=6.13.0"
   },
   "scripts": {
+    "seed": "node database/seed.js",
+    "drop:tables": "node database/dropTables.js",
     "start": "node server/server.js",
     "start:dev": "nodemon server/server.js",
-    "drop:tables": "node database/dropTables.js",
-    "seed": "node database/seed.js",
     "build:dev": "webpack -d --watch",
     "test": "jest",
     "test:server": "jest server",
